refactor(database): move db connection from the application to the library
