+++ angular.json
@@ -21,9 +21,7 @@
               "apps/sample-platform/src/favicon.ico",
               "apps/sample-platform/src/assets"
             ],
-            "styles": [
-              "apps/sample-platform/src/styles.css"
-            ],
+            "styles": ["apps/sample-platform/src/styles.css"],
             "scripts": []
           },
           "configurations": {
@@ -82,10 +80,7 @@
               "apps/sample-platform/tsconfig.app.json",
               "apps/sample-platform/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**",
-              "!apps/sample-platform/**"
-            ]
+            "exclude": ["**/node_modules/**", "!apps/sample-platform/**"]
           }
         },
         "test": {
@@ -120,13 +115,8 @@
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
-            "tsConfig": [
-              "apps/sample-platform-e2e/tsconfig.e2e.json"
-            ],
-            "exclude": [
-              "**/node_modules/**",
-              "!apps/sample-platform-e2e/**"
-            ]
+            "tsConfig": ["apps/sample-platform-e2e/tsconfig.e2e.json"],
+            "exclude": ["**/node_modules/**", "!apps/sample-platform-e2e/**"]
           }
         }
       }
@@ -144,9 +134,7 @@
             "outputPath": "dist/apps/api",
             "main": "apps/api/src/main.ts",
             "tsConfig": "apps/api/tsconfig.app.json",
-            "assets": [
-              "apps/api/src/assets"
-            ]
+            "assets": ["apps/api/src/assets"]
           },
           "configurations": {
             "production": {
@@ -175,10 +163,7 @@
               "apps/api/tsconfig.app.json",
               "apps/api/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**",
-              "!apps/api/**"
-            ]
+            "exclude": ["**/node_modules/**", "!apps/api/**"]
           }
         },
         "test": {
@@ -318,10 +303,7 @@
               "libs/testing/tsconfig.lib.json",
               "libs/testing/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**",
-              "!libs/testing/**"
-            ]
+            "exclude": ["**/node_modules/**", "!libs/testing/**"]
           }
         },
         "test": {
@@ -378,10 +360,7 @@
               "libs/api-implementation/xml/tsconfig.lib.json",
               "libs/api-implementation/xml/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**",
-              "!libs/api-implementation/xml/**"
-            ]
+            "exclude": ["**/node_modules/**", "!libs/api-implementation/xml/**"]
           }
         },
         "test": {
@@ -407,9 +386,7 @@
             "outputPath": "dist/apps/database",
             "main": "apps/database/src/main.ts",
             "tsConfig": "apps/database/tsconfig.app.json",
-            "assets": [
-              "apps/database/src/assets"
-            ]
+            "assets": ["apps/database/src/assets"]
           },
           "configurations": {
             "production": {
@@ -438,10 +415,7 @@
               "apps/database/tsconfig.app.json",
               "apps/database/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**",
-              "!apps/database/**"
-            ]
+            "exclude": ["**/node_modules/**", "!apps/database/**"]
           }
         },
         "test": {
@@ -449,6 +423,35 @@
           "options": {
             "jestConfig": "apps/database/jest.config.js",
             "tsConfig": "apps/database/tsconfig.spec.json",
+            "passWithNoTests": true
+          }
+        }
+      }
+    },
+    "api-implementation-database": {
+      "root": "libs/api-implementation/database",
+      "sourceRoot": "libs/api-implementation/database/src",
+      "projectType": "library",
+      "schematics": {},
+      "architect": {
+        "lint": {
+          "builder": "@angular-devkit/build-angular:tslint",
+          "options": {
+            "tsConfig": [
+              "libs/api-implementation/database/tsconfig.lib.json",
+              "libs/api-implementation/database/tsconfig.spec.json"
+            ],
+            "exclude": [
+              "**/node_modules/**",
+              "!libs/api-implementation/database/**"
+            ]
+          }
+        },
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "options": {
+            "jestConfig": "libs/api-implementation/database/jest.config.js",
+            "tsConfig": "libs/api-implementation/database/tsconfig.spec.json",
             "passWithNoTests": true
           }
         }
