+++ apps/api/src/main.ts
@@ -6,8 +6,6 @@
 import { Logger } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 
-import { connect } from '../../database/src/main'
-
 import { AppModule } from './app/app.module';
 
 async function bootstrap() {
@@ -15,8 +13,6 @@ async function bootstrap() {
   const globalPrefix = 'api';
   app.setGlobalPrefix(globalPrefix);
   const port = process.env.PORT || 5555;
-
-  connect();
 
   await app.listen(port, () => {
     Logger.log('Listening at http://localhost:' + port + '/' + globalPrefix);
