+++ libs/api-implementation/database/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  name: 'api-implementation-database',
+  preset: '../../../jest.config.js',
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../../coverage/libs/api-implementation/database',
+};
