+++ libs/api-implementation/database/src/lib/api-implementation-database.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { DatabaseConnectionService } from './services/connection.service'
+
+@Module({
+  controllers: [],
+  providers: [DatabaseConnectionService],
+  exports: [],
+})
+export class ApiImplementationDatabaseModule {}
