+++ libs/api-implementation/database/src/lib/util/connection.util.ts
@@ -1,0 +1,31 @@
+import * as Mongoose from "mongoose";
+import { AppConfig } from '../../../../../../config'
+
+let database: Mongoose.Connection;
+export const connect = () => {
+  const uri = `${AppConfig.MONGO_CONNECTION_STRING}`;
+  if (database) {
+    return;
+  }
+  Mongoose.connect(uri, {
+    useNewUrlParser: true,
+    useFindAndModify: true,
+    useUnifiedTopology: true,
+    useCreateIndex: true,
+  });
+  database = Mongoose.connection;
+  
+  database.once("open", async () => {
+    console.log("Connected to database");
+  });
+  database.on("error", () => {
+    console.log("Error connecting to database");
+  });
+};
+
+export const disconnect = () => {
+  if (!database) {
+    return;
+  }
+  Mongoose.disconnect();
+};
