+++ libs/api-implementation/database/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../tsconfig.json",
+  "compilerOptions": {
+    "types": ["node", "jest"],
+    "target": "es6"
+  },
+  "include": ["**/*.ts"]
+}
