+++ libs/api-implementation/database/tslint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../../tslint.json",
+  "linterOptions": { "exclude": ["!**/*"] },
+  "rules": {}
+}
